[Refactor] 네이밍 수정 및 import 수정
